chore(DiscordRESTError/flattenErrors): Enhancing readability and performance.
Basically, I used `Object.entries` to iterate over the object's properties `errors` and transform them into an array of `[key, value]`. Then I used reduce to accumulate the error messages into an array, checking whether the current property is a nested object, an array, or an error directly. The `return` in each case returns the existing messages plus the new messages, always prefixing the name of the field in question.