refactor(OggOpusTransformer): readability and performance enhancements
In this new implementation, I created the variable `offset` to control the offset at the current `chunk`. In each iteration of the loop, `this.process(chunk.slice(offset))` is called to process a part of the `chunk` starting from the current `offset`. If the process function returns a value, `this._remainder` is set for the remaining part of the `chunk` and an error event is issued if necessary.

At the end of the `loop`, `this._remainder` is set to null and `cb()` is called to indicate to the stream that it is ready to process the next `chunk`.