refactor(_formatAllowedMentions): modernization
In this new version, I used destructuring assignment to get the `allowedMentions` option from the `this.options` object. I then checked to see if the `allowed` argument was supplied, and returned `allowedMentions` if not. Instead of checking that `allowed.everyone` is true and then adding `"everyone`" to the `result.parse` array, we can simply initialize `result.parse` with `['everyone']` and other values to it as needed. We then check to see if `allowed.roles` or `allowed.users` are true, and if so, we copy up to 100 elements from their respective arrays into `result.roles` or `result.users`. This check prevents the code from trying to copy more than 100 elements, which is the limit for allowed mentions. Finally, we check to see if `allowed.repliedUser` is set, and if it is, we assign its value to `result.replied_user`. And then `result` is returned.

By the way, the error messages triggered to the user are optional, since the problem is already supplied internally because of the code's own behavior, it would only serve for him to have a notion of how it works.