keep the waveform as-is
While currently the docs specify it is a base64 byte array, it might be
more worthwhile to give the API consumer a raw string and let them deal
with it appropriately. Let alone this is probably a more future-proof
variant, as (although very unlikely) the format could change in the future.