chore(BrowserWebSocket): hint arraybuffer preference
Setting `binaryType` on Web API WebSockets looks to be supported all the way since 2012 [1], the Blob path will be kept in case the data still ends up to be a Blob (although this behavior is very unlikely).

In the future, a migration over to Web API WebSockets may be considered, as modern Node.js LTS releases (22+) have support for them.

[1]: https://developer.mozilla.org/en-US/docs/Web/API/WebSocket/binaryType#browser_compatibility